% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_life_history.R
\name{assign_life_history}
\alias{assign_life_history}
\title{assign_life_history}
\usage{
assign_life_history(dat, FishBase = NA, LifeHistoryVars = c("MaxLength",
  "AgeMat", "VonBertK", "Temp", "SpeciesCat", "SpeciesCatName", "b_to_k_ratio"))
}
\arguments{
\item{dat}{the data to be used in the regression etc}

\item{FishBase}{potential fishbase data}

\item{LifeHistoryVars}{}
}
\value{
data frame with life history
}
\description{
\code{assign_life_history} adds in and fills in missing life history information for PRM
}

