% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_phi.R
\name{assign_phi}
\alias{assign_phi}
\title{assign_phi}
\usage{
assign_phi(data, taxdata, phidata, default_phi = 0.188, min_phi = 0.1)
}
\arguments{
\item{data}{data}

\item{taxdata}{the taxdata}

\item{phidata}{the phidata}

\item{default_phi}{default phi}

\item{min_phi}{minimum possible phi}
}
\value{
the phi that produces target BvK ratio
}
\description{
\code{assign_phi} finds the phi in PT model given a BvK ratio
}

