% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_spcategory.R
\name{assign_spcategory}
\alias{assign_spcategory}
\title{assign_spcategory}
\usage{
assign_spcategory(dat, afsismatch = afsis)
}
\arguments{
\item{dat}{the data to assign species categories to}

\item{asfismatch}{ASFIS data to search for matches in}
}
\value{
data frame with assigned species categories
}
\description{
\code{assign_spcategory} adds in and fills in missing life history information for PRM
}

