% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignNearestSpeciesCategory.R
\name{AssignNearestSpeciesCategory}
\alias{AssignNearestSpeciesCategory}
\title{AssignNearestSpeciesCategory}
\usage{
AssignNearestSpeciesCategory(Data, AvailableCategories, AllCategories)
}
\arguments{
\item{Data}{the raw data}

\item{AvailableCategories}{the categories present in the regression model}

\item{AllCategories}{all possible species categories}
}
\value{
nearest neighbot species category
}
\description{
\code{AssignNearestSpeciesCategory} assigns the nearest species category available from regression
}

