% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindResilience.R
\name{FindResilience}
\alias{FindResilience}
\title{Find Resilience}
\usage{
FindResilience(Data)
}
\arguments{
\item{Data}{the data}
}
\value{
Fishbase resilience
}
\description{
Finds Fishbase resilience
}

