% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatForRegression.R
\name{FormatForRegression}
\alias{FormatForRegression}
\title{FormatForRegression}
\usage{
FormatForRegression(f, Data, Fisheries, DependentVariable, CatchVariables,
  CatchLags, LifeHistoryVars, IsLog, IdVar)
}
\arguments{
\item{f}{fishery}

\item{Data}{raw data}

\item{Fisheries}{the total list of all possible fisheries}

\item{DependentVariable}{the left hand variable}

\item{CatchVariables}{vector of catch related variables}

\item{CatchLags}{number of years of catch lags to create}

\item{LifeHistoryVars}{life history variables to include}

\item{IsLog}{is the dependent variable in log space}

\item{IdVar}{the variable marking the id}
}
\value{
formatted data frame
}
\description{
\code{FormatForRegression} takes data and prepares it for PRM regression
}

