% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gum.R
\name{gum}
\alias{gum}
\alias{gum_or}
\alias{gum_and}
\alias{gum_not}
\alias{gum_group}
\alias{gum_phrase}
\title{Wrappers of \code{paste(collapse = )}}
\usage{
gum(...)

gum_or(...)

gum_and(...)

gum_not(...)

gum_group(...)

gum_phrase(...)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}
}
\value{
String.
}
\description{
Wrappers of \code{paste(collapse = )}
}
\examples{
gum("a", "b")
gum_and("a", "b")
gum_or("a", "b")
gum_not("a", "b")
gum_group("a", "b")
gum_phrase("a", "b")
}
