% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gum_query.R
\name{gum_query}
\alias{gum_query}
\title{Expand '?' to 0 or one occurrance of a term}
\usage{
gum_query(x, y, rev = FALSE)
}
\arguments{
\item{x, y}{Character. It can contain one 'word?'.}

\item{rev}{Revert the order of \code{x} and \code{y}.}
}
\value{
Character. A collapsed version of \code{x} and \code{y} where \verb{?} is expanded
to 0 or one occurrance of a term.
}
\description{
Expand '?' to 0 or one occurrance of a term
}
\examples{
gum_query(c("a", "b"), "c?")
}
