% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gum_assessment.R
\name{run_gum_assessment}
\alias{run_gum_assessment}
\title{run_gum_assessment}
\usage{
run_gum_assessment(dat)
}
\arguments{
\item{dat}{data frame with at minimum colnames Year,Catch, SciName}
}
\value{
results a data frame with assessment results
}
\description{
Run assessment part of global upside model (GUM)
}

