% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_post_prm_pt_cmsy.R
\name{run_post_prm_pt_cmsy}
\alias{run_post_prm_pt_cmsy}
\title{Run Post PRM Pella-Tomlinson parameterized CatchMSY}
\usage{
run_post_prm_pt_cmsy(dat, b_to_k_ratio = 0.4, res = "Medium",
  start_bio = NA, mid_bio = NA, final_bio = NA, CommonError = 0.05,
  sigR = 0, Smooth = F, Display = F, n = 10000, NumCPUs = 1)
}
\arguments{
\item{dat}{the catch data and life history information}

\item{b_to_k_ratio}{ratio of biomass at bmsy to biomass at K}

\item{res}{resilience of the species}

\item{CommonError}{default sigma of b/bmsy}

\item{sigR}{standard deviation of recruitment deviates}

\item{Smooth}{T/F to smooth catch histories with running median}

\item{Display}{T/F to show outputs as they are created}

\item{n}{number of iterations to try}

\item{NumCPUs}{number of CPUs to use for parallel processing}
}
\value{
output of CatchMSY
}
\description{
Run Post PRM Pella-Tomlinson parameterized CatchMSY
}

