% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunProjections.R
\name{RunProjection}
\alias{RunProjection}
\title{RunProjections}
\usage{
RunProjection(Data, BaselineYear, NumCPUs, StatusQuoPolicy = "StatusQuoA",
  Policies = c("StatusQuoOpenAccess", "Opt", "CatchShare",
  "StatusQuoFForever", "StatusQuoBForever", "Fmsy", "CloseDown"),
  Discount = 0, CatchShareCost = 0.77, CatchSharePrice = 1.31,
  IdVar = "IdOrig", bvec = seq(1e-08, 2.5, length.out = 30), tol = 0.1,
  beta = 1.3, ProjectionTime = 30, default_catch_share = 0)
}
\arguments{
\item{ProjectionTime}{}
}
\description{
\code{RunProjections} is a function for running projections based on pella tomlinson parameters
}

