% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnowProjections.R
\name{SnowProjections}
\alias{SnowProjections}
\title{Run Snow Projections}
\usage{
SnowProjections(s, Data, BaselineYear, Stocks, IdVar, bvec, Discount, tol, beta,
  CatchSharePrice, CatchShareCost, Policies, ProjectionTime, TempStockMatrix,
  StatusQuoPolicy)
}
\arguments{
\item{StatusQuoPolicy}{}
}
\description{
Run Snow Projections
}

